// dashboard/src/components/project-details/tabs/project-services/_generated/getProjectServicesQuery.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
};

export type AuthResponse = {
  __typename?: 'AuthResponse';
  accessToken: Scalars['String']['output'];
};

/** The status of a container */
export enum ContainerStatus {
  Pending = 'PENDING',
  Running = 'RUNNING',
  Stopped = 'STOPPED'
}

export type CreateProjectInput = {
  description: Scalars['String']['input'];
  name: Scalars['String']['input'];
};

export type CreateServiceInput = {
  config: DeploymentConfigInput;
  name: Scalars['String']['input'];
  projectId: Scalars['String']['input'];
  type: ServiceType;
};

export type Deployment = {
  __typename?: 'Deployment';
  _id: Scalars['ID']['output'];
  config: DeploymentConfig;
  containerId: Maybe<Scalars['String']['output']>;
  containerStatus: Maybe<ContainerStatus>;
  createdAt: Scalars['DateTime']['output'];
  serviceId: Scalars['String']['output'];
  status: DeploymentStatus;
  updatedAt: Scalars['DateTime']['output'];
};

export type DeploymentConfig = {
  __typename?: 'DeploymentConfig';
  cpuLimit: Maybe<Scalars['Int']['output']>;
  image: Scalars['String']['output'];
  memoryLimit: Maybe<Scalars['Int']['output']>;
};
// ...
export enum DeploymentStatus {
  Active = 'ACTIVE',
  Draft = 'DRAFT',
  Finished = 'FINISHED'
}
// ...
export type Project = {
  __typename?: 'Project';
  _id: Scalars['ID']['output'];
  createdAt: Scalars['DateTime']['output'];
  description: Scalars['String']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTime']['output'];
};

export type Query = {
  __typename?: 'Query';
  getNodes: Array<Server>;
  getProjectById: Project;
  getProjectServices: Array<ServiceWithDeployment>;
  getProjects: Array<Project>;
  getServiceById: ServiceWithDeployment;
  getServiceDeployments: Array<Deployment>;
  getUsers: Array<User>;
  isSwarmInitialized: Scalars['Boolean']['output'];
};


export type QueryGetProjectByIdArgs = {
  id: Scalars['String']['input'];
};


export type QueryGetProjectServicesArgs = {
  projectId: Scalars['String']['input'];
};


export type QueryGetServiceByIdArgs = {
  id: Scalars['String']['input'];
};


export type QueryGetServiceDeploymentsArgs = {
  serviceId: Scalars['String']['input'];
};

export type Server = {
  __typename?: 'Server';
  _id: Scalars['ID']['output'];
  createdAt: Scalars['DateTime']['output'];
  ip: Scalars['String']['output'];
  name: Scalars['String']['output'];
  nodeId: Scalars['String']['output'];
  status: Scalars['String']['output'];
  updatedAt: Scalars['DateTime']['output'];
};

/** The type of a service */
export enum ServiceType {
  Docker = 'DOCKER'
}

export type ServiceWithDeployment = {
  __typename?: 'ServiceWithDeployment';
  _id: Scalars['ID']['output'];
  activeDeployment: Maybe<Deployment>;
  createdAt: Scalars['DateTime']['output'];
  draftDeployment: Maybe<Deployment>;
  name: Scalars['String']['output'];
  projectId: Scalars['String']['output'];
  type: ServiceType;
  updatedAt: Scalars['DateTime']['output'];
};

export type UpdateProjectInput = {
  description: Scalars['String']['input'];
  name: Scalars['String']['input'];
};

export type User = {
  __typename?: 'User';
  _id: Scalars['ID']['output'];
  email: Scalars['String']['output'];
  password: Scalars['String']['output'];
  username: Scalars['String']['output'];
};

export type GetProjectServicesQueryVariables = Exact<{
  projectId: Scalars['String']['input'];
}>;


export type GetProjectServicesQuery = { __typename?: 'Query', services: Array<{ __typename?: 'ServiceWithDeployment', _id: string, name: string, type: ServiceType, projectId: string, createdAt: any, updatedAt: any, activeDeployment: { __typename?: 'Deployment', status: DeploymentStatus, containerStatus: ContainerStatus | null, createdAt: any, updatedAt: any } | null, draftDeployment: { __typename?: 'Deployment', status: DeploymentStatus, containerStatus: ContainerStatus | null, createdAt: any, updatedAt: any } | null }> };